moved run to the library
