+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod bootloaders;
 pub mod error;
 pub mod hints;
+pub mod run;
 pub mod tasks;
 
 #[cfg(test)]
